Make APIs more consistent
This commit introduce no new code, it simply renames traits and butchering methods.
The goal here is to follow more closely the standard library naming conventions, so
that new users are more likely to find what they need.

The FlattenCow trait has been renamed AsDerefCow, and its method is now named
as_deref, so that it is similar to the naming of the standard library [1]. Similarly,
the flatten butchering method has been renamed to as_deref.

The UnnestCow trait has been renamed to FlattenCow, so that it is similar to the
flatten method in the standard library [2].

[1]: https://doc.rust-lang.org/std/option/enum.Option.html#method.as_deref
[2]: https://doc.rust-lang.org/std/option/enum.Option.html#method.flatten
